Changes: balanced push-pull architecture
At the beginning the idea was to allow the user to generate digital outputs with different formats.. (5, 12V..). In order
to achieve the result I used a customized push-pull structure with three NPN transistors. But that solution wastes a bit of
current and it not a good one, at the end. So, I have changed to a more common push-pull (PNP + NPN), but the output is not
connected to the transistors emitter, but to their collector, so the transistors work always on saturation area.