New JSON messages for the test-configuration Rest-API
The original idea was to allow the test-client to send all bits-configurations trought the Rest-API. But it is not a good idea,
because the ESP32 has no enough memory to store temporary those data.
So, the new idea is to send the test data in higher level description JSON-messages and to let the ESP32 to convert the smaller
data version in the bits-sequences. In this way ESP32 doesn't need to store data in its internal RAM.