sram_manager module has been disabled
sram_manager module manage the SRAMs pool as a big single fast fifo. In the previouse architecture the test-client software
proveide test-data in raw way (bit per bit) ready to be used by the HILO_engine's MCU. But, because the HILO_interface's MCU
(ESP32) has not enough memory to store all (MBytes) data, the new solution is to provide just data-set definition, and let
the MCU to generate the raw data. For this reason the sram_manager module is no more useful :-(
I hope to port this module to the HILO-engine (STM32-M4) architecture.