Experimental files
About the MD files (README usually), it would be nice to generate the diagrams by a text language. So, I am trying to use
plantUML. It is a great app but it looks like full of bugs, so I don't know if it will be usable. But it is still a great
project